+++ lib/all/setting/view/privacy_screen.dart
